Allow components to describe how best to preview
(previewing on govuk-components)

Most components can be previewed normally, inline within a page, as
they would normally be used in a GOV.UK frontend application.

Some are slightly more complicated, and need to be handled different,
so this adds a new key to the component doc that marks how best to
handle them, eg.

- components that generate non-visible markup, like `<meta>` tags,
  should be previwed as the markup that would be inserted, rather
  than being inserted into the page itself
- components that are unique (like the proposition nav), so showing
  many on one page isn't possible (eg, many propoistion navs being
  moved to the header wouldn't work) and should be displayed only
  one on a page, focusing the non-body content.